Add support for loss functions via `DL_calculation.py`.
This commit adds support for loss functions via DL_calculation.py.
This is done in a rushed approach because of an urgent need for this
functionality, which required reprioritizing development efforts.  The
implementation and required inputs are expected to change as we keep
upgrading `DL_calculation.py` to bring it up to date with the extended
features that were recently added, and considering the ongoing
refinements of those features.