Add documentation pages.
Old docs are stored under `doc_src/`.
There were a few minor updates to those.
Eventually they will be removed.

New doc pages are stored in `doc/`.

Regarding the pelicun framework page:
(doc/source/user_guide/pelicun_framework.rst)

I made some minor revisions. The content in this page is not meant to
be final, it just serves as a colletcion of relevant narrative from
the prior documentation pages, which may be useful when we actually
write that page.